pyproject.toml
[tool.poetry]
name = "django_base"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.10"
django-environ = "^0.11.2"
djangorestframework = "^3.15.1"
drf-spectacular = {extras = ["sidecar"], version = "^0.27.2"}
gunicorn = "^23.0.0"
celery = {extras = ["redis"], version = "^5.4.0"}
hiredis = "^2.3.2"
django-celery-beat = "^2.6.0"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
